requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.16.0
dbt-common==1.25.0
dbt-core==1.9.6
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-protos==1.0.317
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
gensim==4.3.3
greenlet==3.2.3
h11==0.16.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.7.0
msgpack==1.1.0
networkx==3.4.2     # final version that supports Python 3.8-3.10
nltk==3.9.1
numpy==1.26.4
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parsedatetime==2.6
pathspec==0.12.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.33.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.49   # Airflow 2.7 requires <2.0
sqlparse==0.5.3
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.22.0
